Split rule file (1 file per entity type), add celebration 'christmas'.
