+++ grammar/it/src/rules_amount.rs
@@ -1,0 +1,258 @@
+use rustling::*;
+use rustling_ontology_values::dimension::*;
+use rustling_ontology_values::dimension::Precision::*;
+use rustling_ontology_values::helpers;
+
+pub fn rules_percentage(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
+    b.rule_2("<number> per cent",
+        number_check!(),
+        b.reg(r"%|per ?cento?")?,
+        |number, _| Ok(PercentageValue(number.value().value()))
+    );
+    Ok(())
+}
+
+pub fn rules_finance(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
+    b.rule_2("intersect (X cents)",
+             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit != Some("cent")),
+             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit == Some("cent")),
+             |a, b| helpers::compose_money(a.value(), b.value()));
+    b.rule_3("intersect (and X cents)",
+             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit != Some("cent")),
+             b.reg(r#"e"#)?,
+             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit == Some("cent")),
+             |a, _, b| helpers::compose_money(&a.value(), &b.value()));
+    b.rule_3("intersect (and X)",
+             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit != Some("cent")),
+             b.reg(r#"e"#)?,
+             number_check!(),
+             |a, _, b| helpers::compose_money_number(&a.value(), &b.value()));
+    b.rule_2("intersect",
+             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit != Some("cent")),
+             number_check!(),
+             |a, b| helpers::compose_money_number(&a.value(), &b.value()));
+    b.rule_3("intersect",
+             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit != Some("cent")),
+             b.reg(r#"e"#)?,
+             number_check!(),
+             |a, _, b| helpers::compose_money_number(&a.value(), &b.value()));
+    b.rule_1_terminal("$",
+        b.reg(r#"\$|dollar[oi]"#)?,
+        |_| Ok(MoneyUnitValue { unit: Some("$") })
+    );
+    b.rule_1_terminal("EUR",
+        b.reg(r#"€|[e€]ur(?:o?s?|i)"#)?,
+        |_| Ok(MoneyUnitValue { unit: Some("EUR") })
+    );
+    b.rule_1_terminal("£",
+        b.reg(r#"lir[ae]|pound|£"#)?,
+        |_| Ok(MoneyUnitValue { unit: Some("£") })
+    );
+    b.rule_1_terminal("GBP",
+                      b.reg(r#"gbp|(?:lir[ae] )?sterlin[ae](?: britannich?[ae]| ingles[ei])?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("GBP") })
+    );
+    b.rule_1_terminal("USD",
+        b.reg(r#"\$|us[d\$]|dollar[oi]? (?:american[oi]|u\.?s\.?a\.?|statunitens[ei])"#)?,
+        |_| Ok(MoneyUnitValue { unit: Some("USD") })
+    );
+    b.rule_1_terminal("AUD",
+                      b.reg(r#"au[d\$]|dollar[oi] australian[oi]"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("AUD") })
+    );
+    b.rule_1_terminal("CAD",
+                      b.reg(r#"cad|dollar[oi] canades[ei]"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("CAD") })
+    );
+    b.rule_1_terminal("HKD",
+                      b.reg(r#"hkd|dollar[oi] di hong[- ]?kong"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("HKD") })
+    );
+    b.rule_1_terminal("KR",
+                      b.reg(r#"kr|coron[ae]"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("KR") })
+    );
+    b.rule_1_terminal("DKK",
+                      b.reg(r#"dkk|coron[ae] danes[ei]"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("DKK") })
+    );
+    b.rule_1_terminal("NOK",
+                      b.reg(r#"nok|coron[ae] norveges[ei]"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("NOK") })
+    );
+    b.rule_1_terminal("SEK",
+                      b.reg(r#"sek|coron[ae] svedes[ei]"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("SEK") })
+    );
+    b.rule_1_terminal("CHF",
+                      b.reg(r#"chf|franch?[oi] svizzer[oi]"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("CHF") })
+    );
+    // This is not recognized for a very obscure reason
+    b.rule_1_terminal("RUB",
+                      b.reg(r#"rub(?:l[oi])?(?: russ[oi])?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("RUB") })
+    );
+    b.rule_1_terminal("INR",
+                      b.reg(r#"inr|rupi[ae](?: indian[ae])?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("INR") })
+    );
+    b.rule_1_terminal("JPY",
+                      b.reg(r#"jpy|yens?(?: giappones[ei])?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("JPY") })
+    );
+    b.rule_1_terminal("RMB|CNH|CNY",
+                      b.reg(r#"cny|cnh|rmb|yuans?(?: cines[ei])?|renminbis?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("CNY") })
+    );
+    b.rule_1_terminal("¥",
+                      b.reg(r#"¥"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("¥") })
+    );
+    b.rule_1_terminal("KRW",
+                      b.reg(r#"₩|krw|won(?: (?:sud[- ]?)?corean[oi])?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("KRW") })
+    );
+    b.rule_1_terminal("Bitcoin",
+        b.reg(r#"฿|bitcoins?"#)?,
+        |_| Ok(MoneyUnitValue { unit: Some("฿") })
+    );
+    b.rule_1_terminal("cent",
+                      b.reg(r#"cent(?:esim[oi]|s)?"#)?,
+                      |_| Ok(MoneyUnitValue { unit: Some("cent") })
+    );
+    b.rule_2("<amount> <unit>",
+             number_check!(),
+             money_unit!(),
+             |a, b| {
+                 Ok(AmountOfMoneyValue {
+                     value: a.value().value(),
+                     unit: b.value().unit,
+                     ..AmountOfMoneyValue::default()
+                 })
+             });
+    b.rule_2("<unit> <amount>",
+             money_unit!(),
+             number_check!(),
+             |a, b| {
+                 Ok(AmountOfMoneyValue {
+                     value: b.value().value(),
+                     unit: a.value().unit,
+                     ..AmountOfMoneyValue::default()
+                 })
+             });
+    b.rule_3("<amount> of <unit>",
+             integer_check!(|integer: &IntegerValue| !integer.group),
+             b.reg(r#"d[i']"#)?,
+             money_unit!(),
+             |a, _, b| {
+                 Ok(AmountOfMoneyValue {
+                     value: a.value().value as f32,
+                     precision: Exact,
+                     unit: b.value().unit,
+                     ..AmountOfMoneyValue::default()
+                 })
+             });
+    b.rule_2("approx <amount-of-money>",
+             b.reg(r#"verso|interno a|(?:approssim|indic|orient)ativamente|(?:all'in)?circa|quasi|più o meno|pressappoco|suppergiù|grosso modo"#)?,
+             amount_of_money_check!(),
+             |_, a| {
+                 Ok(AmountOfMoneyValue {
+                     precision: Approximate,
+                     ..a.value().clone()
+                 })
+             });
+    b.rule_2("<amount-of-money> approx",
+             amount_of_money_check!(),
+             b.reg(r#"(?:all'in)?circa|più o meno"#)?,
+             |a, _| {
+                 Ok(AmountOfMoneyValue {
+                     precision: Approximate,
+                     ..a.value().clone()
+                 })
+             });
+    b.rule_2("exactly <amount-of-money>",
+             b.reg(r#"(?:esatt|precis)amente"#)?,
+             amount_of_money_check!(),
+             |_, a| {
+                 Ok(AmountOfMoneyValue {
+                     precision: Exact,
+                     ..a.value().clone()
+                 })
+             });
+    b.rule_2("exactly <amount-of-money>",
+             amount_of_money_check!(),
+             b.reg(r#"(?:esatt|precis)amente|(?:precis|esatt)[oiae]"#)?,
+             |a, _| {
+                 Ok(AmountOfMoneyValue {
+                     precision: Exact,
+                     ..a.value().clone()
+                 })
+             }
+    );
+    Ok(())
+}
+
+pub fn rules_temperature(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
+    b.rule_1("number as temp",
+             number_check!(),
+             |a| {
+        Ok(TemperatureValue {
+            value: a.value().value(),
+            unit: None,
+            latent: true,
+        })
+    });
+    // FIXME: should be with number check
+    b.rule_2("<latent temp> degrees",
+             temperature_check!(),
+             b.reg(r#"(?:grad[oi]?)|°"#)?,
+             |a, _| {
+                 Ok(TemperatureValue {
+                     value: a.value().value,
+                     unit: Some("degree"),
+                     latent: false,
+                 })
+             });
+    b.rule_2("<temp> Celsius",
+             temperature_check!(),
+             b.reg(r#"c(?:entigrad[oi]|el[cs]ius|\.)?"#)?,
+             |a, _| {
+                 Ok(TemperatureValue {
+                     value: a.value().value,
+                     unit: Some("celsius"),
+                     latent: false,
+                 })
+             });
+    b.rule_2("<temp> Fahrenheit",
+             temperature_check!(),
+             b.reg(r#"f(?:ah?reh?n(?:h?eit)?)?\.?"#)?,
+             |a, _| {
+                 Ok(TemperatureValue {
+                     value: a.value().value,
+                     unit: Some("fahrenheit"),
+                     latent: false,
+                 })
+             });
+    b.rule_2("<temp> Kelvin",
+             temperature_check!(),
+             b.reg(r#"k(?:elvin|\.)?"#)?,
+             |a, _| {
+                 Ok(TemperatureValue {
+                     value: a.value().value,
+                     unit: Some("kelvin"),
+                     latent: false,
+                 })
+             });
+    b.rule_2("<temp> temp below zero",
+             temperature_check!(|temp: &TemperatureValue| !temp.latent),
+             b.reg(r#"sotto (?:lo )?zero"#)?,
+             |a, _| {
+                 Ok(TemperatureValue {
+                     value: -1.0 * a.value().value,
+                     latent: false,
+                     ..*a.value()
+                 })
+             });
+    Ok(())
+}
