+++ grammar/it/src/rules_datetime.rs
@@ -1,481 +1,8 @@
 use rustling::*;
 use rustling_ontology_values::dimension::*;
-use rustling_ontology_values::dimension::Precision::*;
 use rustling_ontology_values::helpers;
-use rustling_ontology_moment::{Weekday, Grain, PeriodComp, Period};
+use rustling_ontology_moment::{Weekday, Grain};
 
-pub fn rules_percentage(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
-    b.rule_2("<number> per cent",
-        number_check!(),
-        b.reg(r"%|per ?cento?")?,
-        |number, _| Ok(PercentageValue(number.value().value()))
-    );
-    Ok(())
-}
-
-pub fn rules_finance(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
-    b.rule_2("intersect (X cents)",
-             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit != Some("cent")),
-             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit == Some("cent")),
-             |a, b| helpers::compose_money(a.value(), b.value()));
-    b.rule_3("intersect (and X cents)",
-             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit != Some("cent")),
-             b.reg(r#"e"#)?,
-             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit == Some("cent")),
-             |a, _, b| helpers::compose_money(&a.value(), &b.value()));
-    b.rule_3("intersect (and X)",
-             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit != Some("cent")),
-             b.reg(r#"e"#)?,
-             number_check!(),
-             |a, _, b| helpers::compose_money_number(&a.value(), &b.value()));
-    b.rule_2("intersect",
-             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit != Some("cent")),
-             number_check!(),
-             |a, b| helpers::compose_money_number(&a.value(), &b.value()));
-    b.rule_3("intersect",
-             amount_of_money_check!(|money: &AmountOfMoneyValue| money.unit != Some("cent")),
-             b.reg(r#"e"#)?,
-             number_check!(),
-             |a, _, b| helpers::compose_money_number(&a.value(), &b.value()));
-    b.rule_1_terminal("$",
-        b.reg(r#"\$|dollar[oi]"#)?,
-        |_| Ok(MoneyUnitValue { unit: Some("$") })
-    );
-    b.rule_1_terminal("EUR",
-        b.reg(r#"€|[e€]ur(?:o?s?|i)"#)?,
-        |_| Ok(MoneyUnitValue { unit: Some("EUR") })
-    );
-    b.rule_1_terminal("£",
-        b.reg(r#"lir[ae]|pound|£"#)?,
-        |_| Ok(MoneyUnitValue { unit: Some("£") })
-    );
-    b.rule_1_terminal("GBP",
-                      b.reg(r#"gbp|(?:lir[ae] )?sterlin[ae](?: britannich?[ae]| ingles[ei])?"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("GBP") })
-    );
-    b.rule_1_terminal("USD",
-        b.reg(r#"\$|us[d\$]|dollar[oi]? (?:american[oi]|u\.?s\.?a\.?|statunitens[ei])"#)?,
-        |_| Ok(MoneyUnitValue { unit: Some("USD") })
-    );
-    b.rule_1_terminal("AUD",
-                      b.reg(r#"au[d\$]|dollar[oi] australian[oi]"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("AUD") })
-    );
-    b.rule_1_terminal("CAD",
-                      b.reg(r#"cad|dollar[oi] canades[ei]"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("CAD") })
-    );
-    b.rule_1_terminal("HKD",
-                      b.reg(r#"hkd|dollar[oi] di hong[- ]?kong"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("HKD") })
-    );
-    b.rule_1_terminal("KR",
-                      b.reg(r#"kr|coron[ae]"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("KR") })
-    );
-    b.rule_1_terminal("DKK",
-                      b.reg(r#"dkk|coron[ae] danes[ei]"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("DKK") })
-    );
-    b.rule_1_terminal("NOK",
-                      b.reg(r#"nok|coron[ae] norveges[ei]"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("NOK") })
-    );
-    b.rule_1_terminal("SEK",
-                      b.reg(r#"sek|coron[ae] svedes[ei]"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("SEK") })
-    );
-    b.rule_1_terminal("CHF",
-                      b.reg(r#"chf|franch?[oi] svizzer[oi]"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("CHF") })
-    );
-    // This is not recognized for a very obscure reason
-    b.rule_1_terminal("RUB",
-                      b.reg(r#"rub(?:l[oi])?(?: russ[oi])?"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("RUB") })
-    );
-    b.rule_1_terminal("INR",
-                      b.reg(r#"inr|rupi[ae](?: indian[ae])?"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("INR") })
-    );
-    b.rule_1_terminal("JPY",
-                      b.reg(r#"jpy|yens?(?: giappones[ei])?"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("JPY") })
-    );
-    b.rule_1_terminal("RMB|CNH|CNY",
-                      b.reg(r#"cny|cnh|rmb|yuans?(?: cines[ei])?|renminbis?"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("CNY") })
-    );
-    b.rule_1_terminal("¥",
-                      b.reg(r#"¥"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("¥") })
-    );
-    b.rule_1_terminal("KRW",
-                      b.reg(r#"₩|krw|won(?: (?:sud[- ]?)?corean[oi])?"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("KRW") })
-    );
-    b.rule_1_terminal("Bitcoin",
-        b.reg(r#"฿|bitcoins?"#)?,
-        |_| Ok(MoneyUnitValue { unit: Some("฿") })
-    );
-    b.rule_1_terminal("cent",
-                      b.reg(r#"cent(?:esim[oi]|s)?"#)?,
-                      |_| Ok(MoneyUnitValue { unit: Some("cent") })
-    );
-    b.rule_2("<amount> <unit>",
-             number_check!(),
-             money_unit!(),
-             |a, b| {
-                 Ok(AmountOfMoneyValue {
-                     value: a.value().value(),
-                     unit: b.value().unit,
-                     ..AmountOfMoneyValue::default()
-                 })
-             });
-    b.rule_2("<unit> <amount>",
-             money_unit!(),
-             number_check!(),
-             |a, b| {
-                 Ok(AmountOfMoneyValue {
-                     value: b.value().value(),
-                     unit: a.value().unit,
-                     ..AmountOfMoneyValue::default()
-                 })
-             });
-    b.rule_3("<amount> of <unit>",
-             integer_check!(|integer: &IntegerValue| !integer.group),
-             b.reg(r#"d[i']"#)?,
-             money_unit!(),
-             |a, _, b| {
-                 Ok(AmountOfMoneyValue {
-                     value: a.value().value as f32,
-                     precision: Exact,
-                     unit: b.value().unit,
-                     ..AmountOfMoneyValue::default()
-                 })
-             });
-    b.rule_2("approx <amount-of-money>",
-             b.reg(r#"verso|interno a|(?:approssim|indic|orient)ativamente|(?:all'in)?circa|quasi|più o meno|pressappoco|suppergiù|grosso modo"#)?,
-             amount_of_money_check!(),
-             |_, a| {
-                 Ok(AmountOfMoneyValue {
-                     precision: Approximate,
-                     ..a.value().clone()
-                 })
-             });
-    b.rule_2("<amount-of-money> approx",
-             amount_of_money_check!(),
-             b.reg(r#"(?:all'in)?circa|più o meno"#)?,
-             |a, _| {
-                 Ok(AmountOfMoneyValue {
-                     precision: Approximate,
-                     ..a.value().clone()
-                 })
-             });
-    b.rule_2("exactly <amount-of-money>",
-             b.reg(r#"(?:esatt|precis)amente"#)?,
-             amount_of_money_check!(),
-             |_, a| {
-                 Ok(AmountOfMoneyValue {
-                     precision: Exact,
-                     ..a.value().clone()
-                 })
-             });
-    b.rule_2("exactly <amount-of-money>",
-             amount_of_money_check!(),
-             b.reg(r#"(?:esatt|precis)amente|(?:precis|esatt)[oiae]"#)?,
-             |a, _| {
-                 Ok(AmountOfMoneyValue {
-                     precision: Exact,
-                     ..a.value().clone()
-                 })
-             }
-    );
-    Ok(())
-}
-
-pub fn rules_duration(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
-    // Basic duration units
-    b.rule_1_terminal("second (unit-of-duration)",
-                      b.reg(r#"sec(?:ond[oi])?"#)?,
-                      |_| Ok(UnitOfDurationValue::new(Grain::Second))
-    );
-    b.rule_1_terminal("minute (unit-of-duration)",
-                      b.reg(r#"min(?:ut[oi])?"#)?,
-                      |_| Ok(UnitOfDurationValue::new(Grain::Minute))
-    );
-    b.rule_1_terminal("hour (unit-of-duration)",
-                      b.reg(r#"or[ae]"#)?,
-                      |_| Ok(UnitOfDurationValue::new(Grain::Hour))
-    );
-    b.rule_1_terminal("day (unit-of-duration)",
-                      b.reg(r#"giorn(?:[oi]|ata)"#)?,
-                      |_| Ok(UnitOfDurationValue::new(Grain::Day))
-    );
-    b.rule_1_terminal("week (unit-of-duration)",
-                      b.reg(r#"settiman[ae]"#)?,
-                      |_| Ok(UnitOfDurationValue::new(Grain::Week))
-    );
-    b.rule_1_terminal("month (unit-of-duration)",
-                      b.reg(r#"mes(?:e|i)"#)?,
-                      |_| Ok(UnitOfDurationValue::new(Grain::Month))
-    );
-    b.rule_1_terminal("year (unit-of-duration)",
-                      b.reg(r#"ann[oi]"#)?,
-                      |_| Ok(UnitOfDurationValue::new(Grain::Year))
-    );
-    b.rule_1_terminal("1 quarter of an hour",
-                      b.reg(r#"(1/4|(?:un|1) quarto) d'ora"#)?,
-                      |_| Ok(DurationValue::new(PeriodComp::minutes(15).into()))
-    );
-    b.rule_1_terminal("1 half an hour",
-                      b.reg(r#"1/2 ora|(?:una |1 )?mezz'?ora"#)?,
-                      |_| Ok(DurationValue::new(PeriodComp::minutes(30).into()))
-    );
-    b.rule_1_terminal("3 quarters of an hour",
-                      b.reg(r#"(?:3/4|tre quarti) d'ora"#)?,
-                      |_| Ok(DurationValue::new(PeriodComp::minutes(45).into()))
-    );
-    // N duration units
-    b.rule_2("<integer> <unit-of-duration>",
-             integer_check_by_range!(0),
-             unit_of_duration_check!(),
-             |integer, uod| Ok(DurationValue::new(PeriodComp::new(uod.value().grain, integer.value().value).into()))
-    );
-    b.rule_3("<number> hours <number>",
-             integer_check_by_range!(0),
-             b.reg(r#"or[ae] e|h"#)?,
-             integer_check_by_range!(0,59),
-             |hour, _, minute| {
-                 let hour_period = Period::from(PeriodComp::new(Grain::Hour, hour.value().clone().value));
-                 let minute_period = Period::from(PeriodComp::new(Grain::Minute, minute.value().clone().value));
-                 Ok(DurationValue::new(hour_period + minute_period))
-             }
-    );
-    b.rule_3("<integer> <unit-of-duration> and a quarter",
-             integer_check_by_range!(0),
-             unit_of_duration_check!(),
-             b.reg(r#"e un quarto"#)?,
-             |integer, uod, _| {
-                 let quarter_period: Period = uod.value().grain.quarter_period().map(|a| a.into()).ok_or_else(|| RuleError::Invalid)?;
-                 Ok(DurationValue::new(quarter_period + PeriodComp::new(uod.value().grain, integer.value().value)))
-             }
-    );
-    b.rule_3("<integer> <unit-of-duration> and a half",
-             integer_check_by_range!(0),
-             unit_of_duration_check!(),
-             b.reg(r#"e mezz[oa]"#)?,
-             |integer, uod, _| {
-                 let half_period: Period = uod.value().grain.half_period().map(|a| a.into()).ok_or_else(|| RuleError::Invalid)?;
-                 Ok(DurationValue::new(half_period + PeriodComp::new(uod.value().grain, integer.value().value)))
-             }
-    );
-    b.rule_3("<integer> <unit-of-duration> and 3 quarters of an hour",
-             integer_check_by_range!(0),
-             unit_of_duration_check!(),
-             b.reg(r#"e tre quarti"#)?,
-             |integer, uod, _| {
-                 let half_period: Period = uod.value().grain.half_period().map(|a| a.into()).ok_or_else(|| RuleError::Invalid)?;
-                 Ok(DurationValue::new(half_period + PeriodComp::new(uod.value().grain, integer.value().value)))
-             }
-    );
-    // Duration combinations
-    b.rule_3("<duration> and <duration>",
-             duration_check!(|duration: &DurationValue| !duration.suffixed),
-             b.reg(r#"e"#)?,
-             duration_check!(|duration: &DurationValue| !duration.prefixed),
-             |a, _, b| Ok(a.value() + b.value())
-    );
-    b.rule_2("<duration> <duration>",
-             duration_check!(|duration: &DurationValue| !duration.suffixed),
-             duration_check!(|duration: &DurationValue| !duration.prefixed),
-             |a, b| Ok(a.value() + b.value())
-    );
-    b.rule_2("<duration> <integer>",
-             duration_check!(|duration: &DurationValue| !duration.prefixed),
-             integer_check_by_range!(0),
-             |duration, integer| helpers::compose_duration_with_integer(duration.value(), integer.value())
-    );
-    // Durations with modifiers / timeline positioning
-    b.rule_2("in-future <duration> (French 'dans 2 mois')",
-             b.reg(r#"[tf]ra"#)?,
-             duration_check!(),
-             |_, duration| duration.value().in_present()
-    );
-    b.rule_2("<duration> later",
-             duration_check!(),
-             b.reg(r"(dopo|più tardi)")?,
-             |duration, _| duration.value().in_present()
-    );
-    b.rule_2("approx <duration>",
-             b.reg(r#"verso|interno a|(?:approssim|indic|orient)ativamente|(?:all'in)?circa|più o meno|pressappoco|suppergiù|grosso modo"#)?,
-             duration_check!(),
-             |_, duration| Ok(duration.value().clone().precision(Precision::Approximate))
-    );
-    b.rule_2("<duration> approx",
-             duration_check!(),
-             b.reg(r#"(?:all'in)?circa|più o meno"#)?,
-             |duration, _| Ok(duration.value().clone().precision(Precision::Approximate))
-    );
-    b.rule_2("exactly <duration>",
-             b.reg(r#"esattamente"#)?,
-             duration_check!(),
-             |_, duration| Ok(duration.value().clone().precision(Precision::Exact))
-    );
-    b.rule_2("<duration> exactly",
-             duration_check!(),
-             b.reg(r#"(?:esatt|precis)(?:[aoie]|amente)"#)?,
-             |duration, _| Ok(duration.value().clone().precision(Precision::Exact))
-    );
-    b.rule_2("during <duration>",
-             b.reg(r#"(?:durante|per)"#)?,
-             duration_check!(),
-             |_, duration| Ok(duration.value().clone().prefixed())
-    );
-    b.rule_2("<duration> ago",
-             duration_check!(),
-             b.reg(r#"fa"#)?,
-             |duration, _| duration.value().ago()
-    );
-    b.rule_2("since <duration>",
-             b.reg(r#"da(?: |l(?:l['oaie])?)"#)?,
-             duration_check!(),
-             |_, duration| {
-                 duration.value().ago()?
-                     .span_to(&helpers::cycle_nth(Grain::Second, 0)?, false)
-             });
-    b.rule_3("<duration> after <datetime>",
-             duration_check!(),
-             b.reg(r#"dopo"#)?,
-             datetime_check!(),
-             |duration, _, datetime| duration.value().after(datetime.value())
-    );
-    b.rule_3("<duration> before <datetime>",
-             duration_check!(),
-             b.reg(r#"prima"#)?,
-             datetime_check!(),
-             |duration, _, datetime| duration.value().after(datetime.value())
-    );
-
-    Ok(())
-}
-
-pub fn rules_cycle(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
-    // Cycle units
-    b.rule_1_terminal("second (cycle)",
-                      b.reg(r#"second[oi]"#)?,
-                      |_| CycleValue::new(Grain::Second)
-    );
-    b.rule_1_terminal("minute (cycle)",
-                      b.reg(r#"minut[oi]"#)?,
-                      |_| CycleValue::new(Grain::Minute)
-    );
-    b.rule_1_terminal("hour (cycle)",
-                      b.reg(r#"or[ae]"#)?,
-                      |_| CycleValue::new(Grain::Hour)
-    );
-    b.rule_1_terminal("day (cycle)",
-                      b.reg(r#"giorn[oi]"#)?,
-                      |_| CycleValue::new(Grain::Day)
-    );
-    b.rule_1_terminal("week (cycle)",
-                      b.reg(r#"settiman[ae]"#)?,
-                      |_| CycleValue::new(Grain::Week)
-    );
-    b.rule_1("month (cycle)",
-             b.reg(r#"mes[ei]"#)?,
-             |_| CycleValue::new(Grain::Month)
-    );
-    b.rule_1("year (cycle)",
-             b.reg(r#"ann[oi]"#)?,
-             |_| CycleValue::new(Grain::Year)
-    );
-    // Cycles with modifiers / timeline positioning
-    b.rule_2("this / in the <cycle>",
-             b.reg(r#"quest[oa']|in"#)?,
-             cycle_check!(),
-             |_, cycle| helpers::cycle_nth(cycle.value().grain, 0)
-    );
-    b.rule_2("<cycle> last",
-             b.reg(r#"scors[oa]|passat[oa]"#)?,
-             cycle_check!(),
-             |_, cycle| helpers::cycle_nth(cycle.value().grain, -1)
-    );
-    b.rule_2("last <cycle>",
-             cycle_check!(),
-             b.reg(r#"scors[oa]|passat[oa]"#)?,
-             |cycle, _| helpers::cycle_nth(cycle.value().grain, -1)
-    );
-    b.rule_2("the next <cycle>",
-             b.reg(r#"(?:il |la )prossim[oa]"#)?,
-             cycle_check!(),
-             |_, cycle| helpers::cycle_nth(cycle.value().grain, 1)
-    );
-    b.rule_2("next <cycle>",
-             b.reg(r#"prossim[oa]"#)?,
-             cycle_check!(),
-             |_, cycle| helpers::cycle_nth(cycle.value().grain, 1)
-    );
-    b.rule_2("<cycle> next",
-             cycle_check!(),
-             b.reg(r#"prossim[oa]|seguent[ei]|che viene|dopo|successiv[oa]"#)?,
-             |cycle, _| helpers::cycle_nth(cycle.value().grain, 1)
-    );
-    b.rule_3("the <cycle> next",
-             b.reg(r#"il|l['ao]"#)?,
-             cycle_check!(),
-             b.reg(r#"prossim[oa]|seguent[ei]|che viene|dopo|successiv[oa]"#)?,
-             |_, cycle, _| helpers::cycle_nth(cycle.value().grain, 1)
-    );
-    b.rule_3("n <cycle> before",
-             integer_check_by_range!(2, 9999),
-             cycle_check!(),
-             b.reg(r#"prima"#)?,
-             |integer, cycle, _| helpers::cycle_nth(cycle.value().grain, -1 * integer.value().value)
-    );
-    b.rule_3("n <cycle> after",
-             integer_check_by_range!(2, 9999),
-             cycle_check!(),
-             b.reg(r#"dopo"#)?,
-             |integer, cycle, _| helpers::cycle_nth(cycle.value().grain, integer.value().value)
-    );
-    // TODO: more <cycle> combinations with N + past/future
-    // LATER
-    // END TODO
-    b.rule_4("<ordinal> <cycle> of <datetime>",
-             ordinal_check_by_range!(1, 9999),
-             cycle_check!(),
-             b.reg(r#"d(?:['i]|el(?:l['ao])?)"#)?,
-             datetime_check!(),
-             |ordinal, cycle, _, datetime| helpers::cycle_nth_after_not_immediate(cycle.value().grain, ordinal.value().value - 1, datetime.value())
-    );
-    b.rule_5("the <ordinal> <cycle> of <datetime>",
-             b.reg(r#"il|l['ao]"#)?,
-             ordinal_check_by_range!(1, 9999),
-             cycle_check!(),
-             b.reg(r#"d(?:['i]|el(?:l['ao])?)"#)?,
-             datetime_check!(),
-             |_, ordinal, cycle, _, datetime| helpers::cycle_nth_after_not_immediate(cycle.value().grain, ordinal.value().value - 1, datetime.value())
-    );
-    b.rule_4("the <cycle> of <datetime>",
-             b.reg(r#"il|l['ao]"#)?,
-             cycle_check!(),
-             b.reg(r#"d(?:['i]|el(?:l['ao])?)"#)?,
-             datetime_check!(),
-             |_, cycle, _, datetime| helpers::cycle_nth_after_not_immediate(cycle.value().grain, 0, datetime.value())
-    );
-    b.rule_2("the day after the <datetime>",
-             b.reg(r#"(?:l'indomani|il giorno dopo) (?:di|del(?:l[ao'])?)"#)?,
-             datetime_check!(),
-             |_, datetime| helpers::cycle_nth_after_not_immediate(Grain::Day, 1, datetime.value())
-    );
-    b.rule_2("the day before the <datetime>",
-             b.reg(r#"(la vigilia|il giorno prima) (di|del(l(a|o|'))?)"#)?,
-             datetime_check!(),
-             |_, datetime| helpers::cycle_nth_after_not_immediate(Grain::Day, -1, datetime.value())
-    );
-    Ok(())
-}
 
 pub fn rules_datetime(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
     // Basic
@@ -687,6 +214,16 @@ pub fn rules_datetime(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
              b.reg(r#"d(?:['i]|el(?:l['ao])?)"#)?,
              datetime_check!(),
              |_, dow, _, datetime| dow.value().last_of(datetime.value())
+    );
+    b.rule_2("the day after the <datetime>",
+             b.reg(r#"(?:l'indomani|il giorno dopo) (?:di|del(?:l[ao'])?)"#)?,
+             datetime_check!(),
+             |_, datetime| helpers::cycle_nth_after_not_immediate(Grain::Day, 1, datetime.value())
+    );
+    b.rule_2("the day before the <datetime>",
+             b.reg(r#"(la vigilia|il giorno prima) (di|del(l(a|o|'))?)"#)?,
+             datetime_check!(),
+             |_, datetime| helpers::cycle_nth_after_not_immediate(Grain::Day, -1, datetime.value())
     );
     b.rule_4("last <day-of-week> of <datetime> (latent)",
              b.reg(r#"ultim[oa]"#)?,
@@ -1651,15 +1188,6 @@ pub fn rules_datetime(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
              datetime_check!(form!(Form::TimeOfDay(_))),
              |_, a, _, b| a.value().smart_span_to(b.value(), false)
     );
-    b.rule_2("by/before <duration> (French: 'd ici'",
-             b.reg(r#"prima|entro"#)?,
-             duration_check!(),
-             |_, duration| {
-                 let start = helpers::cycle_nth(Grain::Second, 0)?;
-                 let end = duration.value().in_present()?;
-                 start.span_to(&end, false)
-             }
-    );
     b.rule_2("before <time-of-day>",
              b.reg(r#"prima|entro"#)?,
              datetime_check!(),
@@ -1693,415 +1221,129 @@ pub fn rules_datetime(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
     Ok(())
 }
 
-pub fn rules_temperature(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
-    b.rule_1("number as temp",
-             number_check!(),
-             |a| {
-        Ok(TemperatureValue {
-            value: a.value().value(),
-            unit: None,
-            latent: true,
-        })
-    });
-    // FIXME: should be with number check
-    b.rule_2("<latent temp> degrees",
-             temperature_check!(),
-             b.reg(r#"(?:grad[oi]?)|°"#)?,
-             |a, _| {
-                 Ok(TemperatureValue {
-                     value: a.value().value,
-                     unit: Some("degree"),
-                     latent: false,
-                 })
-             });
-    b.rule_2("<temp> Celsius",
-             temperature_check!(),
-             b.reg(r#"c(?:entigrad[oi]|el[cs]ius|\.)?"#)?,
-             |a, _| {
-                 Ok(TemperatureValue {
-                     value: a.value().value,
-                     unit: Some("celsius"),
-                     latent: false,
-                 })
-             });
-    b.rule_2("<temp> Fahrenheit",
-             temperature_check!(),
-             b.reg(r#"f(?:ah?reh?n(?:h?eit)?)?\.?"#)?,
-             |a, _| {
-                 Ok(TemperatureValue {
-                     value: a.value().value,
-                     unit: Some("fahrenheit"),
-                     latent: false,
-                 })
-             });
-    b.rule_2("<temp> Kelvin",
-             temperature_check!(),
-             b.reg(r#"k(?:elvin|\.)?"#)?,
-             |a, _| {
-                 Ok(TemperatureValue {
-                     value: a.value().value,
-                     unit: Some("kelvin"),
-                     latent: false,
-                 })
-             });
-    b.rule_2("<temp> temp below zero",
-             temperature_check!(|temp: &TemperatureValue| !temp.latent),
-             b.reg(r#"sotto (?:lo )?zero"#)?,
-             |a, _| {
-                 Ok(TemperatureValue {
-                     value: -1.0 * a.value().value,
-                     latent: false,
-                     ..*a.value()
-                 })
-             });
+pub fn rules_datetime_with_duration(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
+    b.rule_2("by/before <duration> (French: 'd ici'",
+             b.reg(r#"prima|entro"#)?,
+             duration_check!(),
+             |_, duration| {
+                 let start = helpers::cycle_nth(Grain::Second, 0)?;
+                 let end = duration.value().in_present()?;
+                 start.span_to(&end, false)
+             }
+    );
+
+    Ok(())
+
+}
+
+pub fn rules_datetime_with_nth_cycle(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
+    b.rule_1("month (cycle)",
+             b.reg(r#"mes[ei]"#)?,
+             |_| CycleValue::new(Grain::Month)
+    );
+    b.rule_1("year (cycle)",
+             b.reg(r#"ann[oi]"#)?,
+             |_| CycleValue::new(Grain::Year)
+    );
+    // Cycles with modifiers / timeline positioning
+    b.rule_2("this / in the <cycle>",
+             b.reg(r#"quest[oa']|in"#)?,
+             cycle_check!(),
+             |_, cycle| helpers::cycle_nth(cycle.value().grain, 0)
+    );
+    b.rule_2("<cycle> last",
+             b.reg(r#"scors[oa]|passat[oa]"#)?,
+             cycle_check!(),
+             |_, cycle| helpers::cycle_nth(cycle.value().grain, -1)
+    );
+    b.rule_2("last <cycle>",
+             cycle_check!(),
+             b.reg(r#"scors[oa]|passat[oa]"#)?,
+             |cycle, _| helpers::cycle_nth(cycle.value().grain, -1)
+    );
+    b.rule_2("the next <cycle>",
+             b.reg(r#"(?:il |la )prossim[oa]"#)?,
+             cycle_check!(),
+             |_, cycle| helpers::cycle_nth(cycle.value().grain, 1)
+    );
+    b.rule_2("next <cycle>",
+             b.reg(r#"prossim[oa]"#)?,
+             cycle_check!(),
+             |_, cycle| helpers::cycle_nth(cycle.value().grain, 1)
+    );
+    b.rule_2("<cycle> next",
+             cycle_check!(),
+             b.reg(r#"prossim[oa]|seguent[ei]|che viene|dopo|successiv[oa]"#)?,
+             |cycle, _| helpers::cycle_nth(cycle.value().grain, 1)
+    );
+    b.rule_3("the <cycle> next",
+             b.reg(r#"il|l['ao]"#)?,
+             cycle_check!(),
+             b.reg(r#"prossim[oa]|seguent[ei]|che viene|dopo|successiv[oa]"#)?,
+             |_, cycle, _| helpers::cycle_nth(cycle.value().grain, 1)
+    );
+    b.rule_3("n <cycle> before",
+             integer_check_by_range!(2, 9999),
+             cycle_check!(),
+             b.reg(r#"prima"#)?,
+             |integer, cycle, _| helpers::cycle_nth(cycle.value().grain, -1 * integer.value().value)
+    );
+    b.rule_3("n <cycle> after",
+             integer_check_by_range!(2, 9999),
+             cycle_check!(),
+             b.reg(r#"dopo"#)?,
+             |integer, cycle, _| helpers::cycle_nth(cycle.value().grain, integer.value().value)
+    );
+    // TODO: more <cycle> combinations with N + past/future
+    // LATER
+    // END TODO
+    b.rule_4("<ordinal> <cycle> of <datetime>",
+             ordinal_check_by_range!(1, 9999),
+             cycle_check!(),
+             b.reg(r#"d(?:['i]|el(?:l['ao])?)"#)?,
+             datetime_check!(),
+             |ordinal, cycle, _, datetime| helpers::cycle_nth_after_not_immediate(cycle.value().grain, ordinal.value().value - 1, datetime.value())
+    );
+    b.rule_5("the <ordinal> <cycle> of <datetime>",
+             b.reg(r#"il|l['ao]"#)?,
+             ordinal_check_by_range!(1, 9999),
+             cycle_check!(),
+             b.reg(r#"d(?:['i]|el(?:l['ao])?)"#)?,
+             datetime_check!(),
+             |_, ordinal, cycle, _, datetime| helpers::cycle_nth_after_not_immediate(cycle.value().grain, ordinal.value().value - 1, datetime.value())
+    );
+    b.rule_4("the <cycle> of <datetime>",
+             b.reg(r#"il|l['ao]"#)?,
+             cycle_check!(),
+             b.reg(r#"d(?:['i]|el(?:l['ao])?)"#)?,
+             datetime_check!(),
+             |_, cycle, _, datetime| helpers::cycle_nth_after_not_immediate(cycle.value().grain, 0, datetime.value())
+    );
     Ok(())
 }
 
-pub fn rules_numbers(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
-    b.rule_2("intersect",
-             number_check!(|number: &NumberValue| number.grain().unwrap_or(0) > 1),
-             number_check!(),
-             |a, b| helpers::compose_numbers(&a.value(), &b.value()));
-    b.rule_3("intersect with and",
-             number_check!(|number: &NumberValue| number.grain().unwrap_or(0) > 1),
-             b.reg(r#"e"#)?,
-             number_check!(),
-             |a, _, b| helpers::compose_numbers(&a.value(), &b.value()));
-    // Keep the order of patterns as is, otherwise 'undici' is caught with 'un'
-    b.rule_1_terminal("number (0..19)",
-                      b.reg(r#"(dici(?:assette|otto|annove)|(?:un|do|tre|quattor|quin|se)dici|zero|un[oa']?|due|tr[eé]|quattro|cinque|sei|sette|otto|nove|dieci)"#)?,
-                      |text_match| {
-                          let value = match text_match.group(1).as_ref() {
-                              "zero" => 0,
-                              "un" => 1,
-                              "un'" => 1,
-                              "uno" => 1,
-                              "una" => 1,
-                              "due" => 2,
-                              "tre" => 3,
-                              "tré" => 3,
-                              "quattro" => 4,
-                              "cinque" => 5,
-                              "sei" => 6,
-                              "sette" => 7,
-                              "otto" => 8,
-                              "nove" => 9,
-                              "dieci" => 10,
-                              "undici" => 11,
-                              "dodici" => 12,
-                              "tredici" => 13,
-                              "quattordici" => 14,
-                              "quindici" => 15,
-                              "sedici" => 16,
-                              "diciassette" => 17,
-                              "diciotto" => 18,
-                              "diciannove" => 19,
-                              _ => return Err(RuleError::Invalid.into()),
-                          };
-                          IntegerValue::new(value)
-                      }
+
+pub fn rules_cycle(b: &mut RuleSetBuilder<Dimension>) -> RustlingResult<()> {
+    // Cycle units
+    b.rule_1_terminal("second (cycle)",
+                      b.reg(r#"second[oi]"#)?,
+                      |_| CycleValue::new(Grain::Second)
     );
-    b.rule_1_terminal("number (20..90)",
-                      b.reg(r#"(venti|trenta|(?:(?:quar|cinqu|sess|sett|ott|nov)anta))"#)?,
-                      |text_match| {
-                          let value = match text_match.group(1).as_ref() {
-                              "venti" => 20,
-                              "trenta" => 30,
-                              "quaranta" => 40,
-                              "cinquanta" => 50,
-                              "sessanta" => 60,
-                              "settanta" => 70,
-                              "ottanta" => 80,
-                              "novanta" => 90,
-                              _ => return Err(RuleError::Invalid.into()),
-                          };
-                          IntegerValue::new(value)
-                      });
-    b.rule_2("number (21..29 31..39 41..49 51..59 61..69 71..79 81..89 91..99)",
-             b.reg(r#"(venti?|trenta?|(?:(?:quar|cinqu|sess|sett|ott|nov)anta?))"#)?,
-             integer_check_by_range!(1, 9),
-             |text_match, b| {
-                 let value = match text_match.group(1).as_ref() {
-                     "venti" => 20,
-                     "trenta" => 30,
-                     "quaranta" => 40,
-                     "cinquanta" => 50,
-                     "sessanta" => 60,
-                     "settanta" => 70,
-                     "ottanta" => 80,
-                     "novanta" => 90,
-                     "vent" => 20,
-                     "trent" => 30,
-                     "quarant" => 40,
-                     "cinquant" => 50,
-                     "sessant" => 60,
-                     "settant" => 70,
-                     "ottant" => 80,
-                     "novant" => 90,
-                     _ => return Err(RuleError::Invalid.into())
-                 };
-                 IntegerValue::new(value + b.value().value)
-             });
-    b.rule_1_terminal("number 100..1000",
-                      b.reg(r#"(cento?|duecento|trecento|quattrocento|cinquecento|seicento|settecento|ottocento|novecento|mil(?:le|a))"#)?,
-                      |text_match| {
-                          let value = match text_match.group(1).as_ref() {
-                              "cent" => 100,
-                              "cento" => 100,
-                              "duecento" => 200,
-                              "trecento" => 300,
-                              "quattrocento" => 400,
-                              "cinquecento" => 500,
-                              "seicento" => 600,
-                              "settecento" => 700,
-                              "ottocento" => 800,
-                              "novecento" => 900,
-                              "mille" => 1000,
-                              "mila" => 1000,
-                              _ => return Err(RuleError::Invalid.into())
-                          };
-                          IntegerValue::new_with_grain(value, 2)
-                      });
-    b.rule_2("numbers 100..199",
-             integer_check_by_range!(100, 100),
-             integer_check_by_range!(0, 99),
-             |_, b| IntegerValue::new(b.value().value + 100));
-    b.rule_3("numbers 200..999",
-             integer_check_by_range!(2, 9),
-             integer_check_by_range!(100, 100),
-             integer_check_by_range!(0, 99),
-             |a, b, c| IntegerValue::new(a.value().value * b.value().value + c.value().value));
-    b.rule_1_terminal("hundred",
-                      b.reg(r#"cento"#)?,
-                      |_| IntegerValue::new_with_grain(100, 2)
+    b.rule_1_terminal("minute (cycle)",
+                      b.reg(r#"minut[oi]"#)?,
+                      |_| CycleValue::new(Grain::Minute)
     );
-    b.rule_2("N hundreds",
-             integer_check_by_range!(1, 99),
-             b.reg(r#"cento"#)?,
-             |a, _| {
-                 Ok(IntegerValue {
-                     value: a.value().value * 100,
-                     grain: Some(2),
-                     ..IntegerValue::default()
-                 })
-             });
-    b.rule_1_terminal("thousand",
-                      b.reg(r#"mil(?:le|a)"#)?,
-                      |_| IntegerValue::new_with_grain(1000, 3)
+    b.rule_1_terminal("hour (cycle)",
+                      b.reg(r#"or[ae]"#)?,
+                      |_| CycleValue::new(Grain::Hour)
     );
-    b.rule_2("N thousands",
-             integer_check_by_range!(1, 999),
-             b.reg(r#"mil(?:le|a)"#)?,
-             |a, _| {
-                 Ok(IntegerValue {
-                     value: a.value().value * 1000,
-                     grain: Some(3),
-                     ..IntegerValue::default()
-                 })
-             });
-    b.rule_1_terminal("million",
-                      b.reg(r#"milione?"#)?,
-                      |_| IntegerValue::new_with_grain(1000000, 6)
+    b.rule_1_terminal("day (cycle)",
+                      b.reg(r#"giorn[oi]"#)?,
+                      |_| CycleValue::new(Grain::Day)
     );
-    b.rule_2("N millions",
-             integer_check_by_range!(1, 999),
-             b.reg(r#"milion[ei]?(?: e)?"#)?,
-             |a, _| {
-                 Ok(IntegerValue {
-                     value: a.value().value * 1000000,
-                     grain: Some(6),
-                     ..IntegerValue::default()
-                 })
-             });
-    b.rule_1_terminal("billion",
-                      b.reg(r#"miliardo"#)?,
-                      |_| IntegerValue::new_with_grain(1000000000, 9)
+    b.rule_1_terminal("week (cycle)",
+                      b.reg(r#"settiman[ae]"#)?,
+                      |_| CycleValue::new(Grain::Week)
     );
-    b.rule_2("N billions",
-             integer_check_by_range!(1, 999),
-             b.reg(r#"miliard[oi]"#)?,
-             |a, _| {
-                 Ok(IntegerValue {
-                     value: a.value().value * 1000000000,
-                     grain: Some(9),
-                     ..IntegerValue::default()
-                 })
-             });
-    b.rule_1_terminal("integer (numeric)",
-                      b.reg(r#"(\d{1,18})"#)?,
-                      |text_match| {
-                          let value: i64 = text_match.group(1).parse()?;
-                          IntegerValue::new(value)
-                      });
-
-
-
-    b.rule_1_terminal("integer with thousands separator .",
-                      b.reg(r#"(\d{1,3}(\.\d\d\d){1,5})"#)?,
-                      |text_match| {
-                          let reformatted_string = text_match.group(1).replace(".", "");
-                          let value: i64 = reformatted_string.parse()?;
-                          IntegerValue::new(value)
-                      });
-    b.rule_1_terminal("decimal number",
-                      b.reg(r#"(\d*,\d+)"#)?,
-                      |text_match| {
-                          let reformatted_string = text_match.group(1).replace(",", ".");
-                          let value: f32 = reformatted_string.parse()?;
-                          FloatValue::new(value)
-                      });
-    b.rule_3("number dot number",
-             number_check!(|number: &NumberValue| !number.prefixed()),
-             b.reg(r#"punto|virgola"#)?,
-             number_check!(|number: &NumberValue| !number.suffixed()),
-             |a, _, b| {
-                 let power = b.value().value().to_string().chars().count();
-                 let coeff = 10.0_f32.powf(-1.0 * power as f32);
-                 Ok(FloatValue {
-                     value: b.value().value() * coeff + a.value().value(),
-                     ..FloatValue::default()
-                 })
-             });
-
-    b.rule_4("number dot zero ... number",
-             number_check!(|number: &NumberValue| !number.prefixed()),
-             b.reg(r#"punto|virgola"#)?,
-             b.reg(r#"(?:(?:zero )*(?:zero))"#)?,
-             number_check!(|number: &NumberValue| !number.suffixed()),
-             |a, _, zeros, b| {
-                 let power = zeros.group(0).split_whitespace().count() + b.value().value().to_string().chars().count();
-                 let coeff = 10.0_f32.powf(-1.0 * power as f32);
-                 Ok(FloatValue {
-                     value: b.value().value() * coeff + a.value().value(),
-                     ..FloatValue::default()
-                 })
-             });
-
-    b.rule_1_terminal("decimal with thousands separator",
-                      b.reg(r#"(\d+(\.\d\d\d)+,\d+)"#)?,
-                      |text_match| {
-                          let reformatted_string = text_match.group(1).replace(".", "").replace(",", ".");
-                          let value: f32 = reformatted_string.parse()?;
-                          FloatValue::new(value)
-                      });
-    b.rule_2("numbers prefix with -, negative or minus",
-             b.reg(r#"-|meno"#)?,
-             number_check!(|number: &NumberValue| !number.prefixed()),
-             |_, a| -> RuleResult<NumberValue> {
-                 Ok(match a.value().clone() {
-                     // checked
-                     NumberValue::Integer(integer) => {
-                         IntegerValue {
-                             value: integer.value * -1,
-                             prefixed: true,
-                             ..integer
-                         }
-                             .into()
-                     }
-                     NumberValue::Float(float) => {
-                         FloatValue {
-                             value: float.value * -1.0,
-                             prefixed: true,
-                             ..float
-                         }
-                             .into()
-                     }
-                 })
-             });
-    b.rule_2("numbers prefix with +, positive",
-             b.reg(r#"\+"#)?,
-             number_check!(|number: &NumberValue| !number.prefixed()),
-             |_, a| -> RuleResult<NumberValue> {
-                 Ok(match a.value().clone() {
-                     // checked
-                     NumberValue::Integer(integer) => {
-                         IntegerValue {
-                             prefixed: true,
-                             ..integer
-                         }
-                             .into()
-                     }
-                     NumberValue::Float(float) => {
-                         FloatValue {
-                             prefixed: true,
-                             ..float
-                         }
-                             .into()
-                     }
-                 })
-             }
-    );
-    b.rule_2("numbers suffixes (K, M, G)",
-             number_check!(|number: &NumberValue| !number.suffixed()),
-             b.reg_neg_lh(r#"([kmg])"#, r#"^[\W\$€]"#)?,
-             |a, text_match| -> RuleResult<NumberValue> {
-                 let multiplier = match text_match.group(0).as_ref() {
-                     "k" => 1000,
-                     "m" => 1000000,
-                     "g" => 1000000000,
-                     _ => return Err(RuleError::Invalid.into()),
-                 };
-                 Ok(match a.value().clone() { // checked
-                     NumberValue::Integer(integer) => {
-                         IntegerValue {
-                             value: integer.value * multiplier,
-                             suffixed: true,
-                             ..integer
-                         }
-                             .into()
-                     }
-                     NumberValue::Float(float) => {
-                         let product = float.value * (multiplier as f32);
-                         if product.floor() == product {
-                             IntegerValue {
-                                 value: product as i64,
-                                 suffixed: true,
-                                 ..IntegerValue::default()
-                             }
-                                 .into()
-                         } else {
-                             FloatValue {
-                                 value: product,
-                                 suffixed: true,
-                                 ..float
-                             }
-                                 .into()
-                         }
-                     }
-                 })
-             });
-
-
-    // Ordinals
-    b.rule_1_terminal("ordinals (1-2-3 abbrev)",
-                      b.reg(r#"(?:il |la )?(1|2|3)[oa°]"#)?,
-                      |text_match| {
-                          let value = match text_match.group(1).as_ref() {
-                              "1" => 1,
-                              "2" => 2,
-                              "3" => 3,
-                              _ => return Err(RuleError::Invalid.into())
-                          };
-                          Ok(OrdinalValue::new(value))
-                      });
-    b.rule_1_terminal("ordinals (primo..10)",
-                      b.reg(r#"((?:il |la )?1[oa°]|zeresim|prim|second|terz|quart|quint|sest|settim|ottav|non|decim)[oiae]"#)?,
-                      |text_match| {
-                          let value = match text_match.group(1).as_ref() {
-                              "zeresim" => 0,
-                              "prim" => 1,
-                              "second" => 2,
-                              "terz" => 3,
-                              "quart" => 4,
-                              "quint" => 5,
-                              "sest" => 6,
-                              "settim" => 7,
-                              "ottav" => 8,
-                              "non" => 9,
-                              "decim" => 10,
-                              _ => return Err(RuleError::Invalid.into())
-                          };
-                          Ok(OrdinalValue::new(value))
-                      });
     Ok(())
 }
